pyproject.toml
[tool.poetry]
name = "aws_mkdocs"
version = "0.1.0"
description = ""
authors = ["Thomas Setz <thomas@setz.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
constructs = "^10.1.124"
aws-cdk-lib = "^2.45.0"
flake8 = "^5.0.4"


[tool.poetry.group.dev.dependencies]
mkdocs = "^1.4.0"
pytest = "^7.1.3"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
